+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented here. The format is based
 - Shadow rendering is now default-enabled (#489)
 - All mesh geometry that is used by example models is now centralized in the `geometry/` folder in
   OSC's resource directory (previously: some models had their own `Geometry/` dir, #560)
+- The simulation tab' time scrubber is now in the top toolbar, and includes controls for stepping
+  between states and changing playback speed (#563)
 
 ## [0.3.2] - 2023/01/09